+++ ui/components/buttons/RELEASENOTES.md
@@ -11,6 +11,7 @@
 - New Dual Stateful variation, which acts similar to a [Toggle Button](/accessibility/patterns/togglebutton/#site-main-content), by using the `slds-button_dual-stateful` class.
   - This variation toggles between the styling of a neutral button and a brand button.
 - Added `slds-button_stretch` modifier for adjusting any style button to take 100% width of its container
+- Added examples of links that look like buttons (to be used sparingly)
 
 ### Changed